add option to wipe cached artifacts

+++ build.sh
@@ -1,3 +1,8 @@
-rm -rf build
+# Usage: ./build.sh <clean>
+if [[ "$1" == "clean" ]]; then
+  # Wipe cmake cache
+  rm -rf build
+fi
+
 cmake --preset macos
 cmake --build build --preset macos
